% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_data.R
\name{balance_class}
\alias{balance_class}
\title{Balances Classes in Dataset}
\usage{
balance_class(
  data,
  class,
  response,
  method = "under",
  prop = 0.5,
  seed = 628,
  thresh = 20,
  quiet = FALSE
)
}
\arguments{
\item{data}{dataset to be balanced.}

\item{class}{categorical variable in dataset to be balanced by. This is an optional argument.}

\item{response}{response variable in dataset.}

\item{method}{specifies whether undersampling or oversample should be performed. Takes the value
"under" or "over".}

\item{prop}{desired distribution of response per each class.}

\item{seed}{an integer for set.seed}

\item{quiet}{a logical specifying whether output should be printed to the console.}
}
\value{
balanced data.table.
}
\description{
Balances classes in dataset by undersampling or oversampling.
}
\details{
This function balances the classes of a dataset via undersampling or oversampling using
the \code{ROSE} package. The \code{class} argument is optional. If it is not entered, then the entire
dataset will be balanced as a whole.
\dontrun{
if(interactive()){
 # NEED TO CHANGE
 data(property_prices)
 balance_class(data = property_prices, response = sale_price, class = crime_rate)
 }
}
}
