% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanliness_smry.R
\name{cleanliness_smry}
\alias{cleanliness_smry}
\title{Create cleanliness summary of dataset}
\usage{
cleanliness_smry(data, ignore = c(), outlier_sd = 3, quiet = FALSE)
}
\arguments{
\item{data}{dataset to be analysed.}

\item{ignore}{columns in dataset to be ignored.}

\item{outlier_sd}{Number of standard deviations away from the mean for a observation to be
classified as an outlier. Default is 3.}

\item{quiet}{print messages to console.}
}
\value{
a list of various summaries and plots on how the clean the dataset is.
}
\description{
Creates a summary of how clean a dataset is.
}
\details{
Creates a list of various summaries and plots on how clean a dataset is. The function
returns number of; blank observations, NAs, duplicates, outliers as well as number of plots
for the mentioned figures. Outliers will only be assessed for numeric columns. If certain
columns are to be excluded from this analysis, use the \code{ignore} parameter. To prevent messages
from printing to the console, set \code{quiet} to \code{TRUE} (default is \code{FALSE}).
}
\examples{
\dontrun{
if(interactive()){
 data(property_prices)
 cleanliness_smry(property_prices)
 }
}
}
