% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare Dataset for Modelling}
\usage{
data_prep(
  data,
  response,
  intervals = NULL,
  buckets = NULL,
  na_bucket,
  unmatched_bucket,
  trunc_left = FALSE,
  trunc_right = FALSE,
  include_left = TRUE,
  split_or_fold = 1,
  holdout_ratio = 0,
  unique_row = TRUE,
  rm_low_var = FALSE,
  freq_thresh = 95/5,
  impute = FALSE,
  impute_method = "mice",
  pred_ignore = c(),
  impute_ignore = c(),
  rm_outliers = Inf,
  vif_select = FALSE,
  vif_ignore = c(),
  vif_thresh = 5,
  balance = FALSE,
  balance_class,
  balance_method = "under",
  balance_prop = 0.5,
  scale = FALSE,
  seed = 628,
  quiet = FALSE,
  thresh = 10
)
}
\arguments{
\item{data}{dataset to be analysed.}

\item{response}{response variable to be used in modelling.}

\item{intervals}{a list defining the bands for each of the variables.}

\item{buckets}{a list defining the names of the bands for each of the variables.}

\item{na_bucket}{a character or a list defining the bucket name for entries with \code{NA}.}

\item{unmatched_bucket}{a character or a list defining the bucket name for unmatched entries.}

\item{trunc_left}{a logical specifying whether the band to \code{-Inf} should be created.}

\item{trunc_right}{a logical specifying whether the band to \code{Inf} should be created.}

\item{include_left}{a logical specifying if should include the left or right endpoint for each
interval.}

\item{split_or_fold}{a numeric. If between 0 and 1, dataset is split into training and
testing dataset. The number specifies the percentage of rows to be kept for training. If 1,
dataset is not split and all rows kept for training. If an integer greater than 1,
specifies number of folds to divide the dataset into.}

\item{holdout_ratio}{a numeric between 0 and 1. Specifies what percentage of rows in the
original dataset should be used for the holdout dataset.}

\item{unique_row}{a logical. Whether duplicate rows should be deleted or retained.}

\item{rm_low_var}{a logical. Whether variables with low variance should be deleted or
retained.}

\item{freq_thresh}{the cutoff for the ratio of the most common value to the second most
common value in which a variable is removed when \code{rm_low_var = TRUE}.}

\item{impute}{a logical. Whether missing values and outliers should be imputed.}

\item{impute_method}{method of imputation. Possible methods are "mice" or "mean/mode".}

\item{pred_ignore}{columns in dataset to be not used in data imputation process. Only required if \code{method} = "mice".}

\item{impute_ignore}{columns in dataset to be not imputed.}

\item{rm_outliers}{a numeric where values which are \code{rm_outliers} standard deviations
away from the mean will be imputed. Can either be a single number or a vector of numbers for
each column in \code{data} (including variables in \code{impute_ignore}. By default, set to \code{Inf},
hence no outliers are imputed.}

\item{vif_select}{a logical. Whether stepwise VIF selection should be performed.}

\item{vif_ignore}{columns in dataset to be not removed. Only relevant if \code{remove} is
\code{TRUE}.}

\item{vif_thresh}{threshold of VIF for variables to be removed.}

\item{balance}{a logical. Whether the dataset should be balanced.}

\item{balance_class}{categorical variable in dataset to be balanced by. This is an optional
argument.}

\item{balance_method}{specifies whether undersampling or oversample should be performed.
Takes the value "under" or "over".}

\item{balance_prop}{desired distribution of response per each class.}

\item{scale}{a logical specifying whether the numeric variables should be scaled with 0 mean
and 1 standard deviation.}

\item{seed}{seed for \code{set.seed}.}

\item{quiet}{a logical specifying whether messages should be output to the console.}

\item{used}{for determining whether building a regression or classification model. If number
of unique levels in \code{response} is less than \code{thresh}, then classification, otherwise
regression model.}
}
\value{
List containing prepared dataset with various other metrics and summaries depending
on the arguments entered.
}
\description{
Prepares dataset for modelling by cleaning, banding, imputing and by other
options.
}
\details{
The purpose of this function is to provide a general and off-the-shelf process to
quickly prepare raw datasets for modelling. A large amount of flexibility is provided by
the function and has the options to: band variables, impute missing values and outliers,
perform step-wise VIF selection and balance the dataset by class. For further details on
these process and their arguments, see the following functions respectively contained in the
\code{nano} package: band_data, impute, vif_step and balance_data.

This function also provides the option to: split the dataset into k folds, training, testing,
holdout dataset via the \code{split_or_fold} and \code{holdout_ratio} arguments. To split dataset into
training and testing dataset, set \code{split_or_fold} to be a number between 0 and 1. To divide the
dataset into k folds, set \code{split_or_fold} to be an integer greater than 1. If the dataset has
been split into training and testing, or divided into k folds, additionally, a holdout dataset
can be created. This can be done by using the \code{holdout_ratio} argument. Importantly, the
holdout dataset can only be created if the dataset has been split into training and testing, or
into k folds (i.e. \code{split_or_fold} != 1).

Other features available in this function are: remove variables with low variance via the
\code{rm_low_var} argument and scale numeric variables via the \code{scale} argument.
}
\examples{
\dontrun{
if(interactive()){
 data(property_prices)
 data_prep(data = property_prices, 
           split_or_fold = 0.7, 
           holdout_ratio = 0.1, 
           impute = TRUE, 
           vif_select = TRUE,
           quiet = TRUE)
 }
}
}
