% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_grid.R
\name{delete_grid}
\alias{delete_grid}
\title{Delete Grid from Nano Object}
\usage{
delete_grid(nano, model_no)
}
\arguments{
\item{nano}{\code{nano} objected created by \code{create_nano()}.}

\item{model_no}{position of the grid to be deleted.}
}
\value{
a \code{nano} object with the specified grid deleted.
}
\description{
Deletes a specified grid from a \code{nano} objetct.
}
\details{
Deletes a specified grid from a \code{nano} object. When the grid is deleted, all elements associated with the grid will be
deleted as well. I.e. all the\code{model_no}th elements in each of the lists will be deleted. If deleting a grid which is not the
last one, then subsequent elements are moved up the list by one and renamed appropriately. E.g. if there are three grids and the
second grid is deleted, then the third grid will move to the second position and will be renamed from "grid_3" to "grid_2".
}
\examples{
\dontrun{
if(interactive()){
 nano <- delete_grid(nano, 1)
 }
}
}
