% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_constructors.R
\docType{class}
\name{Grid-class}
\alias{Grid-class}
\alias{create_Grid.H2OGrid}
\alias{create_Grid.H2OAutoML}
\alias{create_Grid.Grid}
\alias{create_Grid}
\title{Creating Grid S4 Object}
\usage{
create_Grid.H2OGrid(obj)

create_Grid.H2OAutoML(obj)

create_Grid.Grid(obj)

create_Grid(obj)
}
\arguments{
\item{obj}{A \code{H2OGrid} or \code{H2OAutoML} object.}
}
\value{
a \code{Grid} S4 object.
}
\description{
constructors for \code{Grid} S4 object
}
\details{
Creates \code{Grid} S4 object from a \code{H2OGrid} or \code{H2OAutoML} object. This object
is used as the \code{grid} element in a \code{nano} object. The \code{Grid} object holds useful
information about the input object which will be used by other functions.
}
\examples{
\dontrun{
if(interactive()){
 library(h2o)
 library(nano)
 
 h2o.init()
 
 # import dataset
 data(property_prices)
 train <- as.h2o(property_prices)
 
 # set the response and predictors
 response <- "sale_price"
 var <- setdiff(colnames(property_prices), response)
 
 # build grids
 grid_1 <- h2o.grid(x               = var,
                    y               = response,
                    training_frame  = train,
                    algorithm       = "randomForest",
                    hyper_params    = list(ntrees = 1:2),
                    nfolds          = 3,
                    seed            = 628)
                    
 Grid_1 <- create_Grid(grid_1)

 }
}
}
