% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute Missing Values in Dataset}
\usage{
impute(
  data,
  method = "mice",
  mice_method = NULL,
  pred_ignore = c(),
  impute_ignore = c(),
  pred_matrix = NULL,
  impute_outlier = Inf,
  seed = 628
)
}
\arguments{
\item{data}{dataset to be analysed.}

\item{method}{method of imputation. Possible methods are "mice" or "mean/mode".}

\item{mice_method}{the method of imputation used by the \code{mice} package. Only required if
\code{method} = "mice".}

\item{pred_ignore}{columns in dataset to be not used in data imputation process. Only required if \code{method} = "mice".}

\item{impute_ignore}{columns in dataset to be not imputed.}

\item{pred_matrix}{optional pre-defined prediction matrix to derive the imputed values when
using the "mice" method.}

\item{impute_outlier}{a numeric where values which are \code{impute_outlier} standard deviations
away from the mean will be imputed. Can either be a single number or a vector of numbers for
each column in \code{data} (including variables in \code{impute_ignore}. By default, set to \code{Inf},
hence no outliers are imputed.}

\item{seed}{seed for \code{set.seed}.}
}
\value{
imputed data.table.
}
\description{
Imputes blanks and NAs in dataset by various methods.
}
\details{
Imputes missing values (blanks or NAs) in dataset. There are two possible methods of imputation. If \code{method}
= "mean/mode", then all missing values in numeric variables will be imputed with its mean and all missing values in
factor variables will be imputed with its mode.

If \code{method} = "mice", then imputation will be performed using the \code{mice} package. The imputation method performed by
\code{mice} can be selected using the \code{mice_method} parameter. A single method can be specified, which will be applied to
all eligible variables in the dataset, or a vector of methods for each variable (including variables listed in
\code{impute_ignore}) can be specified. If the \code{mice_method} parameter is not used, then the default methods selected by
\code{mice} will be used. The \code{pred_ignore} parameter is only applicable when using the "mice" method. This specifies
which variables should not be used in the data imputation process.
}
\examples{
\dontrun{
if(interactive()){
 data(property_prices)
 impute(property_prices)
 }
}
}
