% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pattern.R
\name{missing_pattern}
\alias{missing_pattern}
\title{Calculate Missing Pattern in Dataset}
\usage{
missing_pattern(data, ignore = c(), plot = TRUE)
}
\arguments{
\item{data}{dataset to be analysed.}

\item{ignore}{columns in dataset to be ignored.}

\item{plot}{a logical indicating whether the results should be plotted.}
}
\value{
a list of various summaries and plots on the missing pattern.
}
\description{
Creates a summary of number of missing values by combination of variables.
}
\details{
Identifies all combinations of variables which are missing values (NAs or blanks).
Then calculates the frequency of missing values for each identified combination and then
produces a heatmap of the result.
}
\examples{
\dontrun{
if(interactive()){
 data(property_prices)
 miss <- missing_pattern(property_prices)
 # plot heat map
 miss$plot()
 }
}
}
