% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano-package.R
\docType{package}
\name{nano-package}
\alias{nano}
\alias{nano-package}
\title{nano: Data Visualisation and Model Selection}
\description{
Aim of package is to streamline entire modelling process from data preparation to model diagnostics. Has user friendly functions which creates interactive plots, builds wide range of models and allow for easy comparison between various models.
}
\author{
\strong{Maintainer}: Dilshan Wijesena \email{dilshan.wijesena@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
