% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_constructors.R
\name{new_nano}
\alias{new_nano}
\alias{validate_nano}
\alias{create_nano}
\title{Creating Nano Object}
\usage{
new_nano(
  x = list(grid = rep(list(NA), 10), model = rep(list(NA), 10), data =
    rep(data.table(NA), 10), varimp = rep(list(NA), 10), pdp = rep(list(NA), 10),
    interaction = rep(list(NA), 10), n_model = as.integer(0))
)

validate_nano(x)

create_nano(
  grid = rep(list(NA), 10),
  model = rep(list(NA), 10),
  data = rep(data.table(NA), 10),
  varimp = rep(list(NA), 10),
  pdp = rep(list(NA), 10),
  interaction = rep(list(NA), 10),
  n_model = as.integer(length(grid) - sum(sapply(grid, typeof) == "logical"))
)
}
\arguments{
\item{x}{list of lists with elements containing the below parameters.}

\item{grid}{list of grids created by \code{h2o}.}

\item{model}{list of models created by \code{h2o}, belonging to their respective grids.}

\item{data}{list of datasets used to create each model.}

\item{varimp}{list of datasets containing variable importance for each model.}

\item{pdp}{list of datasets containing partial dependencies for each model.}

\item{interaction}{list of datasets containing interactions for each model.}

\item{n_model}{number of created models.}
}
\value{
a \code{nano} object
}
\description{
constructors for \code{nano} object
}
\details{
Creates a \code{nano} objected which consists of a list of list. If no arguments are
supplied, \code{nano} object is created with 10 elements initialised for each list. If supplying
arguments, must supply arguments for \code{grid}, \code{model} and \code{data}. These must be in list format
and must have the same length. If supplying the above arguments, it is optional to include
varimp, pdp and interaction. If not supplied, they will be initialised as NA. When supplying
arguments, extra elements will be initialised so total number of elements for each list is 10.
}
\examples{
\dontrun{
if(interactive()){
 obj <- create_nano()
 }
}
}
