% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_load.R
\name{nano_load}
\alias{nano_load}
\title{Load Nano Object}
\usage{
nano_load(path)
}
\arguments{
\item{path}{a character. Path of the directory where the nano object is saved in.}
}
\value{
a nano object.
}
\description{
Loads nano object which has been saved using the \code{nano_save} function.
}
\details{
A nano object is able to be saved using the \code{nano_save} function. The
\code{nano_load} function is then used to load the data back as a nano object.
}
\examples{
\dontrun{
if(interactive()){
 library(h2o)
 library(nano)
 
 h2o.init()
 
 # import dataset
 data(property_prices)
 
 # create model
 nano <- nano_automl(data     = data, 
                     response = "sale_price")
                     
 # save nano object
 nano_save(nano, tempdir())
 
 # load back nano object
 nano_load(tempdir())
 
 }
}
}
