% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_metrics.R
\name{nano_metrics}
\alias{nano_metrics}
\title{Calculate Metrics of Models}
\usage{
nano_metrics(
  nano,
  model_no = nano$n_model,
  metrics,
  data_type = "train",
  plot = FALSE
)
}
\arguments{
\item{nano}{nano object to containing models to calculate metrics from.}

\item{model_no}{a numeric. Vector of model numbers to calculate metrics from.}

\item{metrics}{a character. Vector of metrics to be calculated.}

\item{data_type}{a character. From which data source to calculate metrics from. Can take values: "train", "test", "cv" or "holdout".}

\item{plot}{a logical. Whether to output plots.}
}
\value{
a data.table of the metrics for the chosen models.
}
\description{
Calculate various metrics from models in \code{nano} object.
}
\details{
This function
}
\examples{
\dontrun{
}
}
