% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_predict.R
\name{nano_predict}
\alias{nano_predict}
\title{Calculate Predictions}
\usage{
nano_predict(nano, model_no, data_type = "train", fold = NULL)
}
\arguments{
\item{nano}{a nano object.}

\item{model_no}{a numeric. Position of model in \code{nano} to calculate predictions off.}

\item{data_type}{a character. Subset of dataset to calculate predictions off.}

\item{fold}{a numeric. Fold number to calculate predictions off.}
}
\value{
a data.table with two columns, actual and predicted.
}
\description{
Calculates predictions from model stored in a nano object.
}
\details{
\code{data_type} can take the values: "train", "test", "holdout". However, these values
must first exist in the "data_id" column of the relevant dataset. To calculate the predictions
of a specific fold (if cross-validation was performed), used the \code{fold} argument. If the
\code{fold} argument is specified, then the \code{data_type} argument will be ignored.
}
\examples{
\dontrun{
if(interactive()){
 library(h2o)
 library(nano)
 
 h2o.init()
 
 # import dataset
 data(property_prices)
 
 # create model
 nano <- nano_automl(data     = data, 
                     response = "sale_price")
                     
 # calculate training predictions
 pred <- nano_predict(nano, 1, "train")
 
 }
}
}
