% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_save.R
\name{nano_save}
\alias{nano_save}
\title{Save Nano Object}
\usage{
nano_save(nano, path)
}
\arguments{
\item{nano}{a nano object.}

\item{path}{a character. Path of the directory to save the nano object. The
directory must be empty prior to saving the nano object.}
}
\description{
Save nano object into a new directory.
}
\details{
A nano object is able to be saved using the \code{nano_save} function. The
\code{nano_load} function is then used to load the data back as a nano object.
}
\examples{
\dontrun{
if(interactive()){
 library(h2o)
 library(nano)
 
 h2o.init()
 
 # import dataset
 data(property_prices)
 
 # create model
 nano <- nano_automl(data     = data, 
                     response = "sale_price")
                     
 # save nano object
 nano_save(nano, tempdir())
 
 # load back nano object
 nano_load(tempdir())
 
 }
}
}
