% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_shutdown.R
\name{nano_shutdown}
\alias{nano_shutdown}
\title{Shutdown to H2O Cluster}
\usage{
nano_shutdown(prompt = TRUE)
}
\arguments{
\item{prompt}{a logical specifying whether to prompt the user before shutting down connection
to H2O cluster.}
}
\description{
Shutdown connection to a H2O cluster.
}
\details{
This function is a wrapper for the \code{h2o.shutdown} function from the \code{H2O} package.
Note, running this function will removal all H2O grids and models built during the session
(hence will invalidate all \code{nano} objects). Save all \code{nano} objects first by using the
\code{nano_save} function before running this function. For more details, see the manual for
\code{h2o.shutdown}.
}
