% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_single_pdp.R
\name{nano_single_pdp}
\alias{nano_single_pdp}
\title{Create PDP for a Single Model}
\usage{
nano_single_pdp(model, data, vars, max_levels = 30)
}
\arguments{
\item{model}{a h2o model.}

\item{data}{a dataset. Dataset used to create \code{model}.}

\item{vars}{vector of characters. Vector containing variables in \code{data} to create pdps.}

\item{max_levels}{a numeric. Maximum number of unique levels to calculate pdp for each
variable.}
}
\value{
a data.tables containing pdps for each variable combined together.
}
\description{
Creates PDPs for variable(s) for a single h2o model.
}
\details{
Creates a pdp for each variable specified in the \code{vars} argument given a h2o
model.

For creating pdps, it is recommended to instead use the \code{nano_pdp} function
which is a wrapper for a series of functions which creates pdps. It is able to create
pdps directly from a nano object, for both single and multi models, and has the option
to return plots of the pdps.
}
