% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_model.R
\name{switch_model}
\alias{switch_model}
\title{Switch Model from Grid in Nano Object}
\usage{
switch_model(
  nano,
  model_id,
  model_no,
  varimp_eval = FALSE,
  pdp_eval = FALSE,
  ice_eval = FALSE,
  interaction_eval = FALSE,
  plot = FALSE
)
}
\arguments{
\item{nano}{\code{nano} objected created by \code{create_nano()}.}

\item{model_id}{model_id of the new model to be switched to}

\item{model_no}{position in the list of models of the model to be replaced.}

\item{varimp_eval}{calculate variable importance for the new model.}

\item{pdp_eval}{calculate partial dependencies for the new model. If \code{TRUE}, then will calculate the partial dependencies for each
of the variables the partial dependencies were calculated for the original model. If the original model did not have any partial
dependencies, then partial dependencies will be calculated for each variable.}

\item{interaction_eval}{calculates the interactions for the new model. Follows same logic as \code{pdp_eval}.}
}
\value{
a \code{nano} object with the original model replaced by the new model specified by the \code{model_id}.
}
\description{
Switches a model with another model in its associated grid.
}
\details{
Switches a model in a \code{nano} object with another model in its associated grid. By default, when a new grid is produced, the
model with the best metric (chosen by the user) is set as the associated model for that grid. Hence, the user may call this function
to select a different model from the built grid instead. The \code{model_no} is the position in the list of models of the model to be
replaced. The \code{model_id} is the model_id of the new model which is to replace the original model. The new model must exist in the
associated grid of the original model.
}
\examples{
\dontrun{
if(interactive()){
 # switches the second model in the list nano$model with the model with model_id "grid_2_model_3"
 nano <- switch_model(nano, 2, "grid_2_model_3")
 }
}
}
