% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy Dataset for General Use}
\usage{
tidy_data(data, thresh = 10, retain_names = FALSE, response, quiet = FALSE)
}
\arguments{
\item{data}{dataset to be tidied.}

\item{thresh}{a numeric. Threshold for the maximum unique levels of a numeric variables to
be converted as factor type.}

\item{retain_names}{a logical. Whether to retain original variable names.}

\item{response}{a numeric. Response variable to be used in modelling.}
}
\value{
A data.table.
}
\description{
Tidies dataset for general use by correcting variable types, variable names and
removing blanks/NAs in response variable.
}
\details{
This function tidies dataset for general use by: correcting variable types,
correcting variable names and removing blanks/NAs in response variable. The function
returns the tidied data as a data.table.

For further data cleaning in preparation for modelling, please see the \code{data_prep}
function which provides more extensive and comprehensive pre-processing.
}
