% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare Dataset for Modelling}
\usage{
data_prep(
  data,
  response,
  intervals = NULL,
  buckets = NULL,
  na_bucket,
  unmatched_bucket,
  trunc_left = FALSE,
  trunc_right = FALSE,
  include_left = TRUE,
  split_ratio = 0.7,
  unique_row = TRUE,
  rm_low_var = FALSE,
  freq_thresh = 95/5,
  impute = FALSE,
  impute_method = "mice",
  pred_ignore = c(),
  impute_ignore = c(),
  rm_outliers = Inf,
  vif_select = FALSE,
  vif_ignore = c(),
  vif_thresh = 5,
  balance = FALSE,
  balance_class,
  balance_method = "under",
  balance_prop = 0.5,
  scale = FALSE,
  seed = 628,
  quiet = FALSE,
  thresh = 10
)
}
\arguments{
\item{data}{dataset to be analysed.}

\item{response}{response variable to be used in modelling.}

\item{intervals}{a list defining the bands for each of the variables.}

\item{buckets}{a list defining the names of the bands for each of the variables.}

\item{na_bucket}{a character or a list defining the bucket name for entries with \code{NA}.}

\item{unmatched_bucket}{a character or a list defining the bucket name for unmatched entries.}

\item{trunc_left}{a logical specifying whether the band to \code{-Inf} should be created.}

\item{trunc_right}{a logical specifying whether the band to \code{Inf} should be created.}

\item{include_left}{a logical specifying if should include the left or right endpoint for each
interval.}

\item{split_ratio}{ratio to split dataset into training and test dataset. Set to 1 to not
split the dataset.}

\item{unique_row}{a logical specifying whether duplicate rows should be deleted or retained.}

\item{rm_low_var}{a logical specifying whether variables with low variance should be
deleted or retained.}

\item{freq_thresh}{the cutoff for the ratio of the most common value to the second most
common value in which a variable is removed when \code{rm_low_var = TRUE}.}

\item{impute}{a logical specifying whether missing values and outliers should be imputed.}

\item{impute_method}{method of imputation. Possible methods are "mice" or "mean/mode".}

\item{pred_ignore}{columns in dataset to be not used in data imputation process. Only required if \code{method} = "mice".}

\item{impute_ignore}{columns in dataset to be not imputed.}

\item{rm_outliers}{a numeric where values which are \code{rm_outliers} standard deviations
away from the mean will be imputed. Can either be a single number or a vector of numbers for
each column in \code{data} (including variables in \code{impute_ignore}. By default, set to \code{Inf},
hence no outliers are imputed.}

\item{vif_ignore}{columns in dataset to be not removed. Only relevant if \code{remove} is
\code{TRUE}.}

\item{vif_thresh}{threshold of VIF for variables to be removed.}

\item{balance}{a logical specifying whether the dataset should be balanced.}

\item{balance_class}{categorical variable in dataset to be balanced by. This is an optional
argument.}

\item{balance_method}{specifies whether undersampling or oversample should be performed. Takes
the value "under" or "over".}

\item{balance_prop}{desired distribution of response per each class.}

\item{scale}{a logical specifying whether the numeric variables should be scaled with 0 mean
and 1 standard deviation.}

\item{seed}{seed for \code{set.seed}.}

\item{quiet}{a logical specifying whether messages should be output to the console.}

\item{used}{for determining whether building a regression or classification model. If number
of unique levels in \code{response} is less than \code{thresh}, then classification, otherwise
regression model.}
}
\value{
imputed data.table.
}
\description{
Prepares dataset for modelling by cleaning, banding, imputing and by other
options.
}
\details{
Imputes missing values (blanks or NAs) in dataset. There are two possible methods of imputation. If \code{method}
= "mean/mode", then all missing values in numeric variables will be imputed with its mean and all missing values in
factor variables will be imputed with its mode.

If \code{method} = "mice", then imputation will be performed using the \code{mice} package. The imputation method performed by
\code{mice} can be selected using the \code{mice_method} parameter. A single method can be specified, which will be applied to
all eligible variables in the dataset, or a vector of methods for each variable (including variables listed in
\code{impute_ignore}) can be specified. If the \code{mice_method} parameter is not used, then the default methods selected by
\code{mice} will be used. The \code{pred_ignore} parameter is only applicable when using the "mice" method. This specifies
which variables should not be used in the data imputation process.
}
\examples{
\dontrun{
if(interactive()){
 data(property_prices)
 impute(property_prices)
 }
}
}
