% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_constructors.R
\name{new_nano}
\alias{new_nano}
\alias{validate_nano}
\alias{create_nano}
\title{Creating Nano Object}
\usage{
new_nano(
  x = list(grid = rep(list(NA), 10), model = rep(list(NA), 10), data =
    rep(data.table(NA), 10), varimp = rep(list(NA), 10), pdp = rep(list(NA), 10),
    interaction = rep(list(NA), 10), n_model = as.integer(0))
)

validate_nano(x)

create_nano(
  grid = rep(list(NA), 10),
  model = rep(list(NA), 10),
  data = rep(data.table(NA), 10),
  varimp = rep(list(NA), 10),
  pdp = rep(list(NA), 10),
  interaction = rep(list(NA), 10),
  n_model = as.integer(length(grid) - sum(sapply(grid, typeof) == "logical"))
)
}
\arguments{
\item{grid}{list of grids created by \code{h2o}.}

\item{model}{list of models created by \code{h2o}, belonging to their respective grids.}

\item{data}{list of datasets used to create each model.}

\item{varimp}{list of datasets containing variable importance for each model.}

\item{pdp}{list of datasets containing partial dependencies for each model.}

\item{interaction}{list of datasets containing interactions for each model.}

\item{n_model}{number of created models.}
}
\value{
a \code{nano} object
}
\description{
constructors for \code{nano} object
}
\details{
Creates a list of list with each initialised with 10 elements by default.
}
\examples{
\dontrun{
if(interactive()){
 obj <- create_nano()
 }
}
}
