% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_init.R
\name{nano_init}
\alias{nano_init}
\title{Connect to H2O Cluster}
\usage{
nano_init(min_mem_size = NULL, max_mem_size = NULL)
}
\arguments{
\item{min_mem_size}{a character string specifying the minimum size, in bytes, of the memory
allocation pool in H2O.}

\item{max_mem_size}{a character string specifying the maximum size, in bytes, of the memory
allocation pool in H2O.}
}
\description{
Starts and connect to a H2O cluster.
}
\details{
This function is a wrapper for the \code{h2o.init} function from the \code{H2O} package. This
function is used to start and connect to a H2O cluster. A connection to a H2O cluster must be
established before any modelling. For more details, see the manual for \code{h2o.init}.
}
